更新: musicplayer 示例脚本 [skip ci]
